requirements.txt
Django>=5.0
gunicorn
whitenoise
python-dotenv
langgraph
langchain
langchain-community
langchain-groq
groq
openai
